requirements.txt
pandas==2.2.2
google-generativeai==0.7.2
python-dotenv==1.0.1
pytest==8.2.0
